+++ apps/client/src/pages/community/community-page.css.ts
@@ -5,22 +5,6 @@ export const container = style({
   flexDirection: 'column',
 });
 
-export const emptyPlaceholder = style({
-  display: 'flex',
-  flex: 1,
-  width: '100%',
-  height: '100%',
-  alignItems: 'center',
-  justifyContent: 'center',
-});
-
-export const mapCommunityListContainer = style({
-  display: 'flex',
-  flexDirection: 'column',
-  padding: '2.4rem 1.6rem 0',
-  gap: '1.6rem',
-});
-
 export const bottomFloating = style({
   position: 'fixed',
   bottom: '2.4rem',
@@ -33,8 +17,4 @@ export const bottomFloating = style({
       transform: 'none',
     },
   },
-});
-
-export const placeholder = style({
-  height: 'calc(100svh - 20rem)',
 });
