+++ apps/client/src/pages/community/community-page.tsx
@@ -4,17 +4,13 @@ import { useNavigate } from 'react-router';
 import { Alert, Floating, Navigation } from '@bds/ui';
 import { Icon } from '@bds/ui/icons';
 
-import DetailComment from '@widgets/community/components/detail-comment/detail-comment';
-import EmptyPlaceholder from '@widgets/community/components/empty-placeholder/empty-placeholder';
+import FeedList from '@widgets/community/components/feed-list/feed-list';
 import { ALERT_CONTENT_BODY } from '@widgets/community/constant/alert-content';
-import { EMPTY_POST } from '@widgets/community/constant/empty-content';
 
 import { COMMUNITY_QUERY_OPTIONS } from '@shared/api/domain/community/queries';
-import { useIntersectionObserver } from '@shared/hooks/use-intersection-observer';
 import { routePath } from '@shared/router/path';
 
 import * as styles from './community-page.css';
-import { virtualRef } from '@widgets/mypage/preview.css';
 
 const CommunityPage = () => {
   const navigate = useNavigate();
@@ -22,12 +18,6 @@ const CommunityPage = () => {
     useInfiniteQuery({
       ...COMMUNITY_QUERY_OPTIONS.POSTS(),
     });
-
-  const feedObserverRef = useIntersectionObserver(() => {
-    if (hasNextPage && !isFetchingNextPage) {
-      fetchNextPage();
-    }
-  }, true);
 
   const onClickWrite = () => {
     navigate(routePath.COMMUNITY_WRITE);
@@ -52,31 +42,13 @@ const CommunityPage = () => {
         alertContents={ALERT_CONTENT_BODY.BODY}
         type="info"
       />
-      <article className={styles.mapCommunityListContainer}>
-        {data?.pages.some((page) => (page?.content ?? []).length > 0) ? (
-          data.pages
-            .flatMap((page) => page?.content ?? [])
-            .map((post) => (
-              <DetailComment
-                key={post.postId}
-                title={post.title}
-                text={post.content}
-                writerNickname={post.writerNickname}
-                createdAt={post.createdAt}
-                commentCount={post.commentCount}
-                profileImageUrl={post.profileImageUrl ?? ''}
-                onClick={() => navigate(`/community/detail/${post.postId}`)}
-              />
-            ))
-        ) : (
-          <div className={styles.placeholder}>
-            <div className={styles.emptyPlaceholder}>
-              <EmptyPlaceholder content={EMPTY_POST} />
-            </div>
-          </div>
-        )}
-        <div ref={feedObserverRef} className={virtualRef} />
-      </article>
+
+      <FeedList
+        data={data}
+        fetchNextPage={fetchNextPage}
+        hasNextPage={hasNextPage}
+        isFetchingNextPage={isFetchingNextPage}
+      />
 
       <div className={styles.bottomFloating}>
         <Floating
